Merge astrid into main

+++ cub_src/raycasting/degree_calc.c
@@ -6,7 +6,7 @@
 /*   By: ageels <ageels@student.codam.nl>             +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/02/28 20:53:40 by ageels        #+#    #+#                 */
-/*   Updated: 2023/03/02 12:46:33 by mforstho      ########   odam.nl         */
+/*   Updated: 2023/03/06 21:21:04 by ageels        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,16 +14,16 @@
 
 //	utils
 
-float	deg_to_rad(int a)
+float	deg_to_rad(float a)
 {
 	return (a * M_PI / 180.0);
 }
 
 float	fix_ang(float a)
 {
-	if (a > 359)
-		a -= 360;
+	if (a > 359.0)
+		a -= 360.0;
 	if (a < 0)
-		a += 360;
+		a += 360.0;
 	return (a);
 }
